+++ src/main/java/insane96mcp/iguanatweaksreborn/IguanaTweaksReborn.java
@@ -41,8 +41,8 @@ public class IguanaTweaksReborn
     public static final Logger LOGGER = LogManager.getLogger();
 
     public IguanaTweaksReborn() {
-        ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, ITCommonConfig.CONFIG_SPEC);
         ModLoadingContext.get().registerConfig(ModConfig.Type.CLIENT, ITClientConfig.CONFIG_SPEC);
+        ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, ITCommonConfig.CONFIG_SPEC);
         MinecraftForge.EVENT_BUS.register(this);
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::commonSetup);
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::clientSetup);