+++ src/main/java/insane96mcp/iguanatweaksreborn/module/Modules.java
@@ -1,6 +1,6 @@
 package insane96mcp.iguanatweaksreborn.module;
 
-import insane96mcp.iguanatweaksreborn.setup.ITClientConfig;
+import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
 import insane96mcp.iguanatweaksreborn.setup.ITCommonConfig;
 import insane96mcp.insanelib.base.Module;
 import net.minecraftforge.fml.config.ModConfig;
@@ -16,7 +16,6 @@ public class Modules {
 	public static Module movement;
 	public static Module sleepRespawn;
 	public static Module stackSize;
-	public static Module client;
 
 	public static void init() {
 		combat = Module.Builder.create(Ids.COMBAT, "Combat", ModConfig.Type.COMMON, ITCommonConfig.builder).build();
@@ -28,7 +27,6 @@ public class Modules {
 		movement = Module.Builder.create(Ids.MOVEMENT, "Movement", ModConfig.Type.COMMON, ITCommonConfig.builder).build();
 		sleepRespawn = Module.Builder.create(Ids.SLEEP_RESPAWN, "Sleep & Respawn", ModConfig.Type.COMMON, ITCommonConfig.builder).build();
 		stackSize = Module.Builder.create(Ids.STACK_SIZE, "Stack Sizes", ModConfig.Type.COMMON, ITCommonConfig.builder).build();
-		client = Module.Builder.create(Ids.CLIENT, "Client", ModConfig.Type.CLIENT, ITClientConfig.builder).build();
 	}
 
 	public static class Ids {
@@ -41,6 +39,5 @@ public class Modules {
 		public static final String MOVEMENT = IguanaTweaksReborn.RESOURCE_PREFIX + "movement";
 		public static final String SLEEP_RESPAWN = IguanaTweaksReborn.RESOURCE_PREFIX + "sleep_respawn";
 		public static final String STACK_SIZE = IguanaTweaksReborn.RESOURCE_PREFIX + "stack_size";
-		public static final String CLIENT = IguanaTweaksReborn.RESOURCE_PREFIX + "client";
 	}
 }