+++ src/main/java/insane96mcp/iguanatweaksreborn/module/client/feature/Light.java
@@ -1,6 +1,6 @@
 package insane96mcp.iguanatweaksreborn.module.client.feature;
 
-import insane96mcp.iguanatweaksreborn.module.Modules;
+import insane96mcp.iguanatweaksreborn.module.ClientModules;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
@@ -8,7 +8,7 @@ import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
 
 @Label(name = "Light", description = "Changes to light")
-@LoadFeature(module = Modules.Ids.CLIENT)
+@LoadFeature(module = ClientModules.Ids.CLIENT)
 public class Light extends Feature {
 
     @Config
@@ -19,8 +19,7 @@ public class Light extends Feature {
         super(module, enabledByDefault, canBeDisabled);
     }
 
-    //TODO
     public static boolean shouldDisableNightVisionFlashing() {
-        return false;/*isEnabled(Light.class) && noNightVisionFlashing;*/
+        return isEnabled(Light.class) && noNightVisionFlashing;
     }
 }