+++ src/main/java/insane96mcp/iguanatweaksreborn/module/client/feature/Sound.java
@@ -1,6 +1,6 @@
 package insane96mcp.iguanatweaksreborn.module.client.feature;
 
-import insane96mcp.iguanatweaksreborn.module.Modules;
+import insane96mcp.iguanatweaksreborn.module.ClientModules;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
@@ -8,7 +8,7 @@ import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
 
 @Label(name = "Sound", description = "Changes to sounds")
-@LoadFeature(module = Modules.Ids.CLIENT)
+@LoadFeature(module = ClientModules.Ids.CLIENT)
 public class Sound extends Feature {
 
     @Config