+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+* Requires InsaneLib 1.7.1
+* Added back Client module
+
 ## Alpha 2.13.0
 * Updated to 1.19.1+
 * Increased unmending cap (~~20~~ -> 25 levels)